fix(node): use typescript directly to emit files rather than calling tsc
