+++ packages/node/src/builders/package/utils/models.ts
@@ -1,0 +1,27 @@
+import { JsonObject } from '@angular-devkit/core';
+
+export interface NodePackageBuilderOptions extends JsonObject {
+  main: string;
+  tsConfig: string;
+  outputPath: string;
+  watch: boolean;
+  sourceMap: boolean;
+  assets: Array<AssetGlob | string>;
+  packageJson: string;
+  updateBuildableProjectDepsInPackageJson?: boolean;
+}
+
+export interface NormalizedBuilderOptions extends NodePackageBuilderOptions {
+  files: Array<FileInputOutput>;
+  normalizedOutputPath: string;
+  relativeMainFileOutput: string;
+}
+
+export type FileInputOutput = {
+  input: string;
+  output: string;
+};
+export type AssetGlob = FileInputOutput & {
+  glob: string;
+  ignore: string[];
+};
