+++ packages/node/src/schematics/library/files/lib/tsconfig.lib.json
@@ -4,7 +4,6 @@
     "module": "commonjs",
     "outDir": "<%= offsetFromRoot %>dist/out-tsc",
     "declaration": true,
-    "rootDir": "./src",
     "types": ["node"]
   },
   "exclude": ["**/*.spec.ts"],
